+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
  - Cache instance list for five minutes before refreshing the list from AWS.
+ - Show instance state in instance selection.
 
 ## v1.3.0
  - Define search query as command line argument.